Changed const variables to static const
Introduced a fast_exit mode to the daemon to support a faster restart

Added a call to Wire.end() to the initialization of the I2C bus. This allows  to re-init the bus at any time with the same call.

After half of the timeout the ATTiny starts to re-initialize the I2C bus every second as long as there is no new communication with the RPi.

The combination of measures should allow for a faster recovery if the I2C bus is blocked